% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled}
\alias{acled}
\title{ACLED conflict data}
\format{
A data frame with 2726 rows and 3 variables:
\describe{
\item{\code{country}}{character The name of the country}
\item{\code{country_code}}{World Bank country code}
\item{\code{year}}{double The year of the demonstration events}
\item{\code{events}}{double The number of demonstration events}
\item{\code{country_name}}{character. Official World Bank country name}
\item{\code{region}}{character. Official World Bank regional classification}
\item{\code{income_group}}{character. Official World Bank income classification}
}
}
\source{
https://acleddata.com/aggregated/number-demonstration-events-country-year
}
\usage{
acled
}
\description{
ACLED demonstration data, aggregated by country and year.
}
\details{
The Armed Conflict Location & Event Data Project (ACLED) is a comprehensive dataset of political violence, protest, and disorder events across the world. This dataset provides the number of demonstration events aggregated by country and year, globally. Each record summarizes the total number of demonstration events for a given country and year. For more information, see the official ACLED codebook: https://acleddata.com/resources/codebooks/
}
\keyword{datasets}
