% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bready}
\alias{bready}
\title{Business Ready Dataset}
\format{
A data frame with 101 rows and 5 columns:
\describe{
\item{economy}{Country name (character)}
\item{country_code}{ISO country code (character)}
\item{pillar_1_regulatory_framework}{Score for regulatory framework (numeric)}
\item{pillar_2_public_services}{Score for public services (numeric)}
\item{pillar_3_operational_efficiency}{Score for operational efficiency (numeric)}
}
}
\source{
World Bank B-READY. \url{https://www.worldbank.org/content/dam/sites/b-ready/documents/excel/B-READY_ALL_DATA_2025.zip}
}
\usage{
bready
}
\description{
Scores for regulatory framework, public services, and operational efficiency by country.
}
\keyword{datasets}
