% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_execution}
\alias{budget_execution}
\title{Budget Execution Rates}
\format{
A tibble with 2368 rows and 3 columns:
\describe{
\item{country_code}{ISO country code (character)}
\item{year}{Year (character)}
\item{budget_execution_rate}{Budget execution rate (numeric)}
}
}
\source{
Internal compilation
}
\usage{
budget_execution
}
\description{
Annual budget execution rates by country.
}
\keyword{datasets}
