% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data360_api.R
\name{get_data360_api}
\alias{get_data360_api}
\title{This function retrieves data from the Data360 API.}
\usage{
get_data360_api(dataset_id, indicator_id, pivot = TRUE)
}
\arguments{
\item{dataset_id}{Unique identifier for the database.}

\item{indicator_id}{Indicator ID. If a vector of indicators is provided, the query retrieves all indicators.}

\item{pivot}{Whether or not to pivot the extracted data. Default is true.}
}
\value{
A tibble containing data, including country codes, country names,
years, and values.
}
\description{
This function retrieves data from the Data360 API.
}
\examples{
\dontrun{
  cpi_data <- get_data360_api("WB_WDI", "WB_WDI_FP_CPI_TOTL")
  head(cpi_data)
}

}
