% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{open_budget}
\alias{open_budget}
\title{Open Budget Survey Scores}
\format{
A tibble with 918 rows and 7 columns:
\describe{
\item{country_code}{ISO country code (character)}
\item{year}{Year (numeric)}
\item{budget_transparency_score}{Budget transparency score (numeric)}
\item{supreme_audit_oversight_score}{Supreme audit institution oversight score (numeric)}
\item{oversight_score}{Overall oversight score (numeric)}
\item{legislative_oversight_score}{Legislative oversight score (numeric)}
\item{public_participation_score}{Public participation score (numeric)}
}
}
\source{
International Budget Partnership
}
\usage{
open_budget
}
\description{
Scores from the Open Budget Survey by country and year.
}
\keyword{datasets}
