% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population, Total}
\format{
A data frame with 17195 rows and 3 variables:
\describe{
\item{\code{country_code}}{character World Bank country code.}
\item{\code{year}}{double Year.}
\item{\code{total_population}}{double Population, total.}
}
}
\usage{
population
}
\description{
Total population is based on the de facto definition of population, which counts all residents regardless of legal status or citizenship. The values shown are midyear estimates.
}
\details{
World Bank Data 360. https://data360.worldbank.org/en/int/indicator/WB_WDI_SP_POP_TOTL
}
\keyword{datasets}
