% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_map}
\alias{wb_map}
\title{World Bank Map}
\format{
A data frame with 526 rows and 2 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{geometry}}{list Geometry of the country boundaries}
}
}
\source{
https://datacatalogfiles.worldbank.org/ddh-published/0038272/5/DR0095369/World\%20Bank\%20Official\%20Boundaries\%20(GeoJSON)/World\%20Bank\%20Official\%20Boundaries\%20-\%20Admin\%200.geojson
}
\usage{
wb_map
}
\description{
A shapefile containing all relevant country boundaries and disputed areas used by the World Bank.
}
\details{
This dataset includes both the official country boundaries and any disputed areas as recognized by the World Bank.
}
\keyword{datasets}
