% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled}
\alias{acled}
\title{ACLED conflict data}
\format{
A data frame with 2726 rows and 3 variables:
\describe{
\item{\code{country}}{character The name of the country}
\item{\code{year}}{double The year of the demonstration events}
\item{\code{events}}{double The number of demonstration events}
}
}
\source{
https://acleddata.com/aggregated/number-demonstration-events-country-year
}
\usage{
acled
}
\description{
ACLED demonstration data, aggregated by country and year.
}
\details{
#' The Armed Conflict Location & Event Data Project (ACLED) is a comprehensive dataset of political violence, protest, and disorder events across the world. This dataset aggregates ACLED event data for Asia at the weekly level, including event types such as battles, protests, riots, and violence against civilians. Each record summarizes the number of events, fatalities, and estimated population exposure for a specific week, country, and administrative region. Geographic centroids are provided for spatial analysis. For more information, see the official ACLED codebook: https://acleddata.com/resources/codebooks/
}
\keyword{datasets}
